Add `Schema` methods for `Row`s and make `write` more permissive
If you need to `Legolas.write` and all you have is the row type
definition you get fro `@row`, you need to extract the schema in order
to pass it to `Legolas.write`. However, there's no particularly
convenient or obvious way to do this. To make this and other such
situations easier, I've made the following changes:

- `Schema` constructor methods for `Row`, `Type{<:Row}`, `Schema`, and
  `Type{<:Schema}`. This allows you to call `Schema` on something which
  has or is itself a schema and get back the schema.
- Remove the type restriction for the schema argument of
  `Legolas.write`, thereby permitting passing a row, row type, schema,
  or schema type (or anything else for which a `Schema` method exists).