Use common methods for Base functions instead of generating (#103)
Currently, methods for `==`, `isequal`, `hash`, and `NamedTuple` are
generated by the `@version` macro for every row type that gets defined.
The bodies of these methods always follow the same patterns, none of
which require any information that's only available in the context where
the definitions are generated. In fact, specific methods don't need to
be generated at all; `AbstractRecord`s can all use the same generic
methods for these functions. This provides the following additional
benefits:
- When many schema versions are defined, this significantly reduces the
  excessive number of redundant methods which make method autocompletion
  effectively useless.
- Defining the generic methods in terms of regular functions rather than
  generating expressions in the macro means that it's easier to reason
  about the methods' behavior and ensure overall consistency. For
  example, moving from generating a chain of `==` and `&&` over a
  record's fields to directly using `all` means that `missing` is
  treated consistently (see issue 101).

Note that the generic method for `hash` as defined here loops over the
fields in reverse order. This is to match the previous `foldr` behavior,
ensuring that hashes don't change with this implementation.