Move ConstructionBase dependency to an extension on 1.9+ (#112)
* Move ConstructionBase dependency to an extension on 1.9+

The dependency on ConstructionBase serves to provide a few convenience
methods for users of packages like Accessors; it doesn't provide any
functionality used directly by Legolas. This makes it a perfect
candidate for a package extension.

Package extensions are new in Julia 1.9, and Legolas currently supports
Julia 1.6. That means that we can't fully remove ConstructionBase as a
direct dependency yet, but we can restructure things to make that as
easy as possible once we drop support for earlier Julia versions.

* Bump Aqua compatibility to 0.8 to fix Julia 1.6 test failure

Requires adding stdlib entries to `[compat]`

* Bump version